requirements.txt
Django==2.2.24
djangorestframework==3.12.4
Pillow==7.2.0
pytesseract==0.3.5
pytz==2021.1
sqlparse==0.4.1
